+ p-value correction, HGT attention pipeline, config warnings, dead code cleanup, documentation

+++ src/models/components/gene_attention_gate.py
@@ -117,7 +117,10 @@ class GeneAttentionGate(nn.Module):
             Gate weights of shape (n_cell_types, n_genes), each row sums to 1
         """
         tau = self._temperature_buf.item()
-        return F.softmax(self.gate_logits / tau, dim=-1)
+        # Promote to float32 for softmax stability under AMP.
+        # At low temperature (tau→0.05), logits/tau amplifies by 20x.
+        # Float16 exp() overflows at ~11.09 — any logit > 0.55 would produce NaN.
+        return F.softmax((self.gate_logits / tau).float(), dim=-1).to(self.gate_logits.dtype)
 
     def get_top_genes_per_cell_type(
         self, k: int = 100, gene_names: list[str] | None = None
